requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor==3.3.2
python-dotenv>=0.19.0
pydantic>=1.8.0
email-validator
python-jose[cryptography]>=3.3.0
passlib
python-multipart>=0.0.5
pymongo==4.6.1
PyJWT>=2.8.0
